Update and rename utility.py to cog_template.py
Converted the unused utility cog into a template cog so users can easily add new ones.
